+++ encodings/fsst/src/lib.rs
@@ -15,6 +15,7 @@ mod array;
 mod canonical;
 mod compress;
 mod compute;
+mod dfa;
 mod kernel;
 mod ops;
 mod rules;
